FIX: Problems have been solved * timeStamp of unixstamp was changed for timeStamp of the video time * timeStamp variable type was changed for int * Information about numbers of events and lastTimeStamp is updated every frame

+++ core/NexPlayer_dll/src/NPTracker.cpp
@@ -8,7 +8,7 @@ NPTracker::~NPTracker()
 {
 }
 
-void NPTracker::OnPlayPause(unsigned long long  currentTimeStamp)
+void NPTracker::OnPlayPause(int currentTimeStamp)
 {
 	_numberPlayPause++;
 	_lastTimeStamp = currentTimeStamp;
@@ -19,11 +19,10 @@ int NPTracker::GetNumberOfEvents()
 	return _numberPlayPause;
 }
 
-long long NPTracker::GetLastTimestamp()
+int NPTracker::GetLastTimestamp()
 {
 	return _lastTimeStamp;
 }
-
 
 //-------------------------------------------------------------//
 EXPORT NPTracker* createTracker()
@@ -36,7 +35,7 @@ EXPORT void freeTracker(NPTracker* instance)
 	delete instance;
 }
 
-EXPORT void onPlayPause(NPTracker* instance, unsigned long long  currentTimeStamp)
+EXPORT void onPlayPause(NPTracker* instance, int currentTimeStamp)
 {
 	instance->OnPlayPause(currentTimeStamp);
 }
@@ -46,7 +45,7 @@ int getNumberPlayPauseEvents(NPTracker* instance)
 	return instance->GetNumberOfEvents();
 }
 
-unsigned long long getLastTimeStamp(NPTracker* instance)
+int getLastTimeStamp(NPTracker* instance)
 {
 	return instance->GetLastTimestamp();
 }
